TST: specify dt64 units (#56217)
* TST: specify dt64 units

* specify dt64 unit

+++ pandas/tests/groupby/test_timegrouper.py
@@ -537,7 +537,9 @@ class TestGroupBy:
         for date in dates:
             result = grouped.get_group(date)
             data = [[df.loc[date, "A"], df.loc[date, "B"]]]
-            expected_index = DatetimeIndex([date], name="date", freq="D")
+            expected_index = DatetimeIndex(
+                [date], name="date", freq="D", dtype=index.dtype
+            )
             expected = DataFrame(data, columns=list("AB"), index=expected_index)
             tm.assert_frame_equal(result, expected)
